feat: Add input name abstraction
Now there is additional abstraction layer in the input manager. With this you able to create 'actions' that will impersonate all denoted keys and in functions for example BindAction you will able to pass 'action' instead to call this function for every key on which you want to call your callback. This feature is dramatically decrease the code length and also make it more clear and clean